Changed message type to 32 bit

+++ export.cpp
@@ -19,6 +19,9 @@
  *
  *
  * $Log$
+ * Revision 1.9  2008/12/19 18:23:05  vfrolov
+ * Changed message type to 32 bit
+ *
  * Revision 1.8  2008/11/27 16:19:16  vfrolov
  * Added pBufAppend()
  *
@@ -71,7 +74,7 @@ static VOID CALLBACK buf_append(BYTE **ppBuf, DWORD offset, const BYTE *pSrc, DW
   BufAppend(ppBuf, offset, pSrc, sizeSrc);
 }
 ///////////////////////////////////////////////////////////////
-static BOOL CALLBACK msg_replace_buf(HUB_MSG *pMsg, WORD type, const BYTE *pSrc, DWORD sizeSrc)
+static BOOL CALLBACK msg_replace_buf(HUB_MSG *pMsg, DWORD type, const BYTE *pSrc, DWORD sizeSrc)
 {
   _ASSERTE((type & HUB_MSG_UNION_TYPES_MASK) == HUB_MSG_UNION_TYPE_BUF);
 
@@ -94,7 +97,7 @@ static BOOL CALLBACK msg_replace_buf(HUB_MSG *pMsg, WORD type, const BYTE *pSrc,
   return TRUE;
 }
 ///////////////////////////////////////////////////////////////
-static HUB_MSG *CALLBACK msg_insert_buf(HUB_MSG *pPrevMsg, WORD type, const BYTE *pSrc, DWORD sizeSrc)
+static HUB_MSG *CALLBACK msg_insert_buf(HUB_MSG *pPrevMsg, DWORD type, const BYTE *pSrc, DWORD sizeSrc)
 {
   _ASSERTE((type & HUB_MSG_UNION_TYPES_MASK) == HUB_MSG_UNION_TYPE_BUF);
 
@@ -129,7 +132,7 @@ static HUB_MSG *CALLBACK msg_insert_buf(HUB_MSG *pPrevMsg, WORD type, const BYTE
   return pMsg;
 }
 ///////////////////////////////////////////////////////////////
-static BOOL CALLBACK msg_replace_val(HUB_MSG *pMsg, WORD type, DWORD val)
+static BOOL CALLBACK msg_replace_val(HUB_MSG *pMsg, DWORD type, DWORD val)
 {
   _ASSERTE((type & HUB_MSG_UNION_TYPES_MASK) == HUB_MSG_UNION_TYPE_VAL);
 
@@ -144,7 +147,7 @@ static BOOL CALLBACK msg_replace_val(HUB_MSG *pMsg, WORD type, DWORD val)
   return TRUE;
 }
 ///////////////////////////////////////////////////////////////
-static HUB_MSG *CALLBACK msg_insert_val(HUB_MSG *pPrevMsg, WORD type, DWORD val)
+static HUB_MSG *CALLBACK msg_insert_val(HUB_MSG *pPrevMsg, DWORD type, DWORD val)
 {
   _ASSERTE((type & HUB_MSG_UNION_TYPES_MASK) == HUB_MSG_UNION_TYPE_VAL);
 
@@ -164,7 +167,7 @@ static HUB_MSG *CALLBACK msg_insert_val(HUB_MSG *pPrevMsg, WORD type, DWORD val)
   return pMsg;
 }
 ///////////////////////////////////////////////////////////////
-static BOOL CALLBACK msg_replace_none(HUB_MSG *pMsg, WORD type)
+static BOOL CALLBACK msg_replace_none(HUB_MSG *pMsg, DWORD type)
 {
   _ASSERTE((type & HUB_MSG_UNION_TYPES_MASK) == HUB_MSG_UNION_TYPE_NONE);
 
@@ -178,7 +181,7 @@ static BOOL CALLBACK msg_replace_none(HUB_MSG *pMsg, WORD type)
   return TRUE;
 }
 ///////////////////////////////////////////////////////////////
-static HUB_MSG *CALLBACK msg_insert_none(HUB_MSG *pPrevMsg, WORD type)
+static HUB_MSG *CALLBACK msg_insert_none(HUB_MSG *pPrevMsg, DWORD type)
 {
   _ASSERTE((type & HUB_MSG_UNION_TYPES_MASK) == HUB_MSG_UNION_TYPE_NONE);
